move contact & social data to common.json & update components
