--- data/contacts.json
@@ -1,8 +1,0 @@
-{
-  "title": "Контакти",
-  "address": {
-    "city": "м. Одеса",
-    "street": "вул. Канатна, 42"
-  },
-  "phone": ["+38 (093) 615 03 01"]
-}
