--- data/socials.json
@@ -1,17 +1,0 @@
-[
-  {
-    "name": "facebook",
-    "link": "https://www.facebook.com/mafclubod",
-    "icon": "Facebook"
-  },
-  {
-    "name": "instagram",
-    "link": "https://www.instagram.com/mafia_club_odesa/",
-    "icon": "Instagram"
-  },
-  {
-    "name": "telegram",
-    "link": "https://t.me/MafiaclubOdessa_Bot",
-    "icon": "Telegram"
-  }
-]
